fix:rename to styles.ts, fix Section in Stub.tsx
