+++ src/components/Stub/styles.ts
